MAPSJS-2783: Move tests in TileCreationTest to DecodedTileHelperTest.
